GCC Compliance & Integrating Visibility Settings
+ GCC issued warnings about a circular include introduced with the last change, so I moved the templated functions in GLCanvas from the header and into the CPP file. I also needed a few explicit template instantiations to coerce the stamping out of the necessary functions.
+ The checkmarks in the Debugging menu are now set based on the settings in the preferences.json file.

+++ source/Scene/treemapAsset.cpp
@@ -500,7 +500,7 @@ namespace Asset
    {
       assert(m_VAO.isCreated());
       assert(m_blockColorBuffer.isCreated());
-      assert(static_cast<int>(node->offsetIntoVBO) < m_blockCount);
+      assert(node->offsetIntoVBO < m_blockCount);
 
       constexpr auto colorTupleSize{ sizeof(QVector3D) };
       const auto offsetIntoColorBuffer = node->offsetIntoVBO * colorTupleSize;
